Added example for the ent guide
The ent guide on the encore.dev website can be difficult to get into due to how specific some of the steps can be. As part of preparing a PR for improving it, I ended up creating a new example project that follows the exact tutorial.

This example shows how to use the ent ORM and could even be used as-is by users wanting to start with an ORM. It also shows the necessary trick of creating an empty migration file.