Add an example for serving static assets with Encore
As part of writing a guide for using Encore as a server for serving web applications with injected data from the backend, I've prepared a small example that would probably work well in this repository. I'm also not 100% sure if this is the right approach for this kind of application, so this PR is probably a good way of valdiating that before I push a PR for the guide.